Add files via upload
These 2 files are being used by a server which has nearly 4k members as of now.

It is a simple bot whose purpose is to maintain clean and civil environment in the server by filtering out words which the server mods don't want in and deleting spammy links and notifying a moderator about the user who did this.

Also has a command to display inspirational quotes by famous people. 

Used flask to deploy bot on uptimerobot.